Rename projects section to your projects section
